requirements.txt
aiogram>=3.4.0
fastapi>=0.110.0
uvicorn>=0.27.0
asyncpg>=0.29.0
PyYAML>=6.0.1
pydantic-settings>=2.2.1
structlog>=23.1.0
